Move `GlowCommand` to `entity` package + group
